pyproject.toml
[project]
name = "factr-dynamixel"
version = "0.1.0"
description = "FACTR low-cost force-feedback teleoperation with Dynamixel servos"
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [
    { name = "Jason Jingzhou Liu", email = "liujason@cmu.edu" },
    { name = "Yulong Li" },
    # 基于原仓库的修改与扩展
    { name = "Gabriel-Ning", email = "guomning@gmail.com" },
]
dependencies = [
    "numpy",
    "mujoco>=3.1.3",
    "pin",
    "PyYAML",
    "ruamel.yaml",
    "pyserial",
    "pyzmq",
    "dynamixel-sdk>=4.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/factr_teleop", "src/python_utils"]
